requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
python-jose==3.3.0
requests==2.31.0
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
pytest==8.0.2
pytest-flask==1.3.0
